refactor: adjust Input component name and migrate it from a class to a CustomInput function
