+++ src/components/CustomInput/CustomInput.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
-import type { InputProps } from './Input';
-import { Input } from './Input';
+import type { InputProps } from './CustomInput';
+import { CustomInput } from './CustomInput';
 import { userSetUp } from '../../test-utils/test-utils';
 
 const mockOnChange = vi.fn();
@@ -13,25 +13,25 @@ const mockData: InputProps[] = [
     type: 'text',
     value: '',
     onChange: mockOnChange,
-    className: 'searchBar',
+    customClass: 'searchBar',
     placeholder: 'Search for character',
   },
 ];
 
-describe('Input', () => {
+describe('CustomInput', () => {
   beforeEach(() => {
     mockOnChange.mockClear();
   });
 
   mockData.forEach((setOfAttributes) => {
     test('input renders successfully with provided attributes', () => {
-      render(<Input {...setOfAttributes} />);
+      render(<CustomInput {...setOfAttributes} />);
       const input = screen.getByRole('textbox');
       expect(input).toBeInTheDocument();
       expect(input).toHaveAttribute('type', setOfAttributes.type);
       expect(input).toHaveValue(setOfAttributes.value);
-      if (setOfAttributes.className) {
-        expect(input).toHaveClass(setOfAttributes.className);
+      if (setOfAttributes.customClass) {
+        expect(input).toHaveClass(setOfAttributes.customClass);
       }
       if (setOfAttributes.placeholder) {
         expect(
@@ -41,7 +41,7 @@ describe('Input', () => {
     });
   });
   test('calls the onChange callback handler', async () => {
-    const { user } = userSetUp(<Input {...mockData[1]} />);
+    const { user } = userSetUp(<CustomInput {...mockData[1]} />);
     const input = screen.getByRole('textbox');
     await user.type(input, 'Morty');
     expect(mockOnChange).toHaveBeenCalled();
